OBD-master/requirements.txt
argon2-cffi==23.1.0
asgiref==3.7.2
beautifulsoup4==4.12.2
cloudinary==1.36.0
django-extensions==3.2.3
django-test-without-migrations==0.6
dj-database-url==2.1.0
dj-static==0.0.6
decorator==5.1.1
Django==4.2.11
gunicorn==21.2.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
Pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2024.1
scout-apm==3.2.0
sqlparse==0.4.4
static3==0.7.0
requests==2.31.0
urllib3<2.0.0
xlwt==1.3.0



